requirements.txt
# ArUco Pose Estimation - Core Dependencies

# Core libraries
numpy>=1.20.0
opencv-python>=4.6.0
opencv-contrib-python>=4.6.0
pyyaml>=5.4
packaging>=20.0

# CLI and advanced features
click>=8.0.0
scipy>=1.7.0

# Development and testing (optional - uncomment if needed)
# pytest>=6.0.0
# pytest-cov>=2.0
# black>=21.0
# flake8>=3.9
